Extract execution-start lifecycle into focused execution_lifecycle module
